Optimizations and code clarity

+++ src/com/esotericsoftware/kryo/io/ByteBufferOutput.java
@@ -651,7 +651,7 @@ public class ByteBufferOutput extends Output {
 
 		varInt |= (0x80 << 24);
 		long varLong = (varInt & 0xFFFFFFFFL);
-		varLong |= (((long)(value & 0x7F)) << 32);
+		varLong |= ((value & 0x7F) << 32);
 
 		value >>>= 7;
 
@@ -665,7 +665,7 @@ public class ByteBufferOutput extends Output {
 		}
 
 		varLong |= (0x80L << 32);
-		varLong |= (((long)(value & 0x7F)) << 40);
+		varLong |= ((value & 0x7F) << 40);
 
 		value >>>= 7;
 
@@ -679,7 +679,7 @@ public class ByteBufferOutput extends Output {
 		}
 
 		varLong |= (0x80L << 40);
-		varLong |= (((long)(value & 0x7F)) << 48);
+		varLong |= ((value & 0x7F) << 48);
 
 		value >>>= 7;
 
@@ -693,7 +693,7 @@ public class ByteBufferOutput extends Output {
 		}
 
 		varLong |= (0x80L << 48);
-		varLong |= (((long)(value & 0x7F)) << 56);
+		varLong |= ((value & 0x7F) << 56);
 
 		value >>>= 7;
